+++ src/Imports/ModelImporter.php
@@ -8,6 +8,7 @@ use Maatwebsite\Excel\Concerns\WithCalculatedFormulas;
 use Maatwebsite\Excel\Concerns\WithColumnLimit;
 use Maatwebsite\Excel\Concerns\WithMapping;
 use Maatwebsite\Excel\Concerns\WithProgressBar;
+use Maatwebsite\Excel\Concerns\WithValidation;
 use Maatwebsite\Excel\Row;
 use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 
@@ -46,6 +47,7 @@ class ModelImporter
         $progessBar       = $import instanceof WithProgressBar;
         $withMapping      = $import instanceof WithMapping;
         $withCalcFormulas = $import instanceof WithCalculatedFormulas;
+        $withValidation   = $import instanceof WithValidation && method_exists($import, 'prepareForValidation');
         $endColumn        = $import instanceof WithColumnLimit ? $import->endColumn() : null;
 
         $this->manager->setRemembersRowNumber(method_exists($import, 'rememberRowNumber'));
@@ -56,6 +58,10 @@ class ModelImporter
 
             $row      = new Row($spreadSheetRow, $headingRow);
             $rowArray = $row->toArray(null, $withCalcFormulas, true, $endColumn);
+
+            if ($withValidation) {
+                $rowArray = $import->prepareForValidation($rowArray, $row->getIndex());
+            }
 
             if ($withMapping) {
                 $rowArray = $import->map($rowArray);