+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Added support for `prepareForValidation` on `WithValidation` concern
 - Added support for `withValidator` on `WithValidation` concern
 
 ## [3.1.23] - 2020-09-29